+++ CHANGELOG.md
@@ -94,6 +94,10 @@ high state of flux, you're at risk of it changing without notice.
     - deprecate `Witherable` in favour of `getWitherable` (@anthonyjoeseph)
   - `Semigroup`
     - deprecate `semigroupVoid`, use `void` module instead.
+  - `State`
+    - deprecate `traverseArrayWithIndex` in favour of `traverseReadonlyArrayWithIndex`
+    - deprecate `traverseArray` in favour of `traverseReadonlyArrayWithIndex`
+    - deprecate `sequenceArray` in favour of `traverseReadonlyArrayWithIndex`
   - `Task`
     - deprecate `traverseArrayWithIndex` in favour of `traverseReadonlyArrayWithIndex`
     - deprecate `traverseArray` in favour of `traverseReadonlyArrayWithIndex`
@@ -318,6 +322,9 @@ high state of flux, you're at risk of it changing without notice.
   - `Set`
     - add `getUnionSemigroup`
     - add `getDifferenceMagma`
+  - `State`
+    - add `traverseReadonlyNonEmptyArrayWithIndex`
+    - add `traverseReadonlyArrayWithIndex`
   - `StateReaderTaskEither`
     - add `fromStateK`
     - add `chainStateK`